+++ lua/pl/compat.lua
@@ -34,15 +34,21 @@ compat.is_windows = compat.dir_separator == '\\'
 
 --- execute a shell command, in a compatible and platform independent way.
 -- This is a compatibility function that returns the same for Lua 5.1 and
--- Lua 5.2, on Unixes as well as Windows.
+-- Lua 5.2+.
+--
+-- NOTE: Windows systems can use signed 32bit integer exitcodes. Posix systems
+-- only use exitcodes 0-255, anything else is undefined.
 -- @param cmd a shell command
 -- @return true if successful
 -- @return actual return code
-function compat.execute (cmd)
-    local res1,_,res3 = os.execute(cmd)
+function compat.execute(cmd)
+    local res1,res2,res3 = os.execute(cmd)
+    if res2 == "No error" and res3 == 0 and compat.is_windows then
+      -- os.execute bug in Lua 5.2+ not reporting -1 properly on Windows
+      res3 = -1
+    end
     if compat.lua51 and not compat.jit52 then
         if compat.is_windows then
-            res1 = res1 > 255 and res1 % 256 or res1
             return res1==0,res1
         else
             res1 = res1 > 255 and res1 / 256 or res1
@@ -50,7 +56,6 @@ function compat.execute (cmd)
         end
     else
         if compat.is_windows then
-            res3 = res3 > 255 and res3 % 256 or res3
             return res3==0,res3
         else
             return not not res1,res3