+++ tests/test-path2.lua
@@ -1,23 +1,39 @@
 local path = require 'pl.path'
 local test = require 'pl.test'
-
-relpath = path.relpath
-
-path = '/a/b/c'
+local asserteq = test.asserteq
 
 function slash (p)
     return (p:gsub('\\','/'))
 end
 
+
+--  path.relpath
+
+
+local testpath = '/a/B/c'
+
 function try (p,r)
-    test.asserteq(slash(relpath(p,path)),r)
+    asserteq(slash(path.relpath(p,testpath)),r)
 end
 
-try('/a/b/c/one.lua','one.lua')
-try('/a/b/c/bonzo/two.lua','bonzo/two.lua')
-try('/a/b/three.lua','../three.lua')
+try('/a/B/c/one.lua','one.lua')
+try('/a/B/c/bonZO/two.lua','bonZO/two.lua')
+try('/a/B/three.lua','../three.lua')
 try('/a/four.lua','../../four.lua')
 try('one.lua','one.lua')
 try('../two.lua','../two.lua')
 
 
+--  path.common_prefix
+
+
+asserteq(slash(path.common_prefix("../anything","../anything/goes")),"../anything")
+asserteq(slash(path.common_prefix("../anything/goes","../anything")),"../anything")
+asserteq(slash(path.common_prefix("../anything/goes","../anything/goes")),"../anything")
+asserteq(slash(path.common_prefix("../anything/","../anything/")),"../anything")
+asserteq(slash(path.common_prefix("../anything","../anything")),"..")
+asserteq(slash(path.common_prefix("/hello/world","/hello/world/filename.doc")),"/hello/world")
+asserteq(slash(path.common_prefix("/hello/filename.doc","/hello/filename.doc")),"/hello")
+if path.is_windows then
+    asserteq(path.common_prefix("c:\\hey\\there","c:\\hey"),"c:\\hey")
+end