+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
   - `utils.raise` changed the global `on_error`-level when passing in bad arguments
   - `utils.writefile` now checks and returns errors when writing
+  - `compat.execute` now handles the Windows exitcode -1 properly
 
 ## 1.6.0 (2018-11-23)