+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.6.1 (UNRELEASED)
+## 1.7.0 (UNRELEASED)
 
 ### New features
 
@@ -15,7 +15,7 @@
   - `utils.quit`: exit message is no longer required, and closes the Lua state (on 5.2+).
   - `utils.assert_arg` and `utils.assert_string`: now return the validated value
   - `pl.compat`: now exports the `jit` and `jit52` flags
-
+  - `pretty.write`: now sorts the output for easier diffs [#293](https://github.com/stevedonovan/Penlight/pull/293)
 
 ### Fixes
 
@@ -31,6 +31,7 @@
     (only impacted Windows)
   - `dir.copyfile`, `dir.movefile`, and `dir.makepath` create the new file/path with
     the requested casing, and no longer force lowercase (only impacted Windows)
+  - added a missing assertion on `path.getmtime` [#291](https://github.com/stevedonovan/Penlight/pull/291)
 
 ## 1.6.0 (2018-11-23)